Move process_legacy_payment code to rest api class
