+++ src/Library.php
@@ -9,10 +9,7 @@ namespace Automattic\WooCommerce\Blocks;
 
 defined( 'ABSPATH' ) || exit;
 
-use Automattic\WooCommerce\Blocks\Payments\PaymentResult;
-use Automattic\WooCommerce\Blocks\Payments\PaymentContext;
 use Automattic\WooCommerce\Blocks\Payments\PaymentMethodRegistry;
-use Automattic\WooCommerce\Blocks\RestApi\StoreApi\Utilities\NoticeHandler;
 
 /**
  * Library class.
@@ -32,7 +29,6 @@ class Library {
 		add_filter( 'woocommerce_register_shop_order_post_statuses', array( __CLASS__, 'register_draft_order_post_status' ) );
 		add_filter( 'woocommerce_valid_order_statuses_for_payment', array( __CLASS__, 'append_draft_order_post_status' ) );
 		add_action( 'woocommerce_cleanup_draft_orders', array( __CLASS__, 'delete_expired_draft_orders' ) );
-		add_action( 'woocommerce_rest_checkout_process_payment_with_context', array( __CLASS__, 'process_legacy_payment' ), 999, 2 );
 	}
 
 	/**
@@ -205,56 +201,5 @@ class Library {
 			AND posts.post_modified <= ( NOW() - INTERVAL 1 DAY )
 			"
 		);
-	}
-
-	/**
-	 * Attempt to process a payment for the checkout API if no payment methods support the
-	 * woocommerce_rest_checkout_process_payment_with_context action.
-	 *
-	 * @param PaymentContext $context Holds context for the payment.
-	 * @param PaymentResult  $result  Result of the payment.
-	 */
-	public static function process_legacy_payment( PaymentContext $context, PaymentResult &$result ) {
-		if ( $result->status ) {
-			return;
-		}
-
-		// phpcs:ignore WordPress.Security.NonceVerification
-		$post_data = $_POST;
-
-		// Set constants.
-		wc_maybe_define_constant( 'WOOCOMMERCE_CHECKOUT', true );
-
-		// Add the payment data from the API to the POST global.
-		$_POST = $context->payment_data;
-
-		// Call the process payment method of the chosen gatway.
-		$payment_method_object = $context->get_payment_method_instance();
-
-		if ( ! $payment_method_object instanceof \WC_Payment_Gateway ) {
-			return;
-		}
-
-		$payment_method_object->validate_fields();
-
-		// If errors were thrown, we need to abort.
-		NoticeHandler::convert_notices_to_exceptions( 'woocommerce_rest_payment_error' );
-
-		// Process Payment.
-		$gateway_result = $payment_method_object->process_payment( $context->order->get_id() );
-
-		// Restore $_POST data.
-		$_POST = $post_data;
-
-		// If `process_payment` added notices, clear them. Notices are not displayed from the API -- payment should fail,
-		// and a generic notice will be shown instead if payment failed.
-		wc_clear_notices();
-
-		// Handle result.
-		$result->set_status( isset( $gateway_result['result'] ) && 'success' === $gateway_result['result'] ? 'success' : 'failure' );
-
-		// set payment_details from result.
-		$result->set_payment_details( array_merge( $result->payment_details, $gateway_result ) );
-		$result->set_redirect_url( $gateway_result['redirect'] );
 	}
 }
